Adiciona validação de Clients e altera importação do Swagger p/ router

+++ src/controllers/Clients/index.js
@@ -1,3 +1,5 @@
+const { validationResult } = require('express-validator');
+
 const clientRepository = require('../../services/Clients');
 const medicalRecordsRepository = require('../../services/MedicalRecords');
 
@@ -19,8 +21,14 @@ async function getById(req, res) {
 }
 async function update(req, res) {
   try {
+    const validationErrors = validationResult(req);
+    if (!validationErrors.isEmpty()) {
+      return res.status(400).json(validationErrors.array());
+    }
     const result = await clientRepository.update(req.params.id, req.body);
-    return res.json(result);
+    return result ? res.json(result) : res.status(404).json(`
+      Cliente com ID ${req.params.id} não existe!
+      `);
   } catch (err) {
     return res.status(400).json(err);
   }
@@ -28,6 +36,11 @@ async function update(req, res) {
 
 async function insert(req, res) {
   try {
+    const validationErrors = validationResult(req);
+    if (!validationErrors.isEmpty()) {
+      return res.status(400).json(validationErrors.array());
+    }
+
     const client = req.body;
     const clientResult = await clientRepository.create(client);
     const medRecordResult = await medicalRecordsRepository.create({ client: clientResult.id });
@@ -40,6 +53,10 @@ async function insert(req, res) {
 
 async function remove(req, res) {
   try {
+    const validationErrors = validationResult(req);
+    if (!validationErrors.isEmpty()) {
+      return res.status(404).json(validationErrors.array());
+    }
     const result = await clientRepository.delete(req.params.id);
     return (result.affected ? res.status(200).json(result) : res.status(404).json(result));
   } catch (err) {
